Bead Counter/pyproject.toml
[project]
name = "bead-counter"
version = "0.1.0"
description = "Bead counting tool for Mixtec codices analysis"
requires-python = ">=3.9"
dependencies = [
    "loguru",
    "matplotlib",
    "numpy",
    "pillow",
    "scikit-learn",
    "seaborn",
    "tomli",
    "torch",
    "torchinfo",
    "torchvision",
    "tqdm",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
